feat: update tutorial dialog with gif navigation and improved close button
This commit introduces several changes to the tutorials in the Pomodoro task app. Firstly, we added 'gif' support for adding images in tutorials for better demonstration of features. Functions have been added to enable navigation of these images using arrow buttons. We have also reimplemented the close button functionality. Now, it prompts users to complete the tutorial before closing the dialog, improving the chances of users to learning the app better.